+++ crates/duty-tracker/Cargo.toml
@@ -18,10 +18,10 @@ rustdoc.all = "warn"
 [dependencies]
 btc-notify.workspace = true
 bitcoin.workspace = true
-futures = "0.3.31"
+futures.workspace = true
 strata-bridge-primitives.workspace = true
 strata-bridge-tx-graph.workspace = true
 strata-btcio.workspace = true
 strata-state.workspace = true
-tokio = "1.43.0"
+tokio.workspace = true
 tokio-stream = "0.1.17"
